refactor(web): events
